tools/requirements.txt
pyexr
numpy
matplotlib
Pillow
bs4
html5lib